# docs/swagger.yaml
definitions:
  errcode.Error:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      err:
        type: error
      msg:
        type: string
    type: object
...
  service.ChapterContentResponse:
    properties:
      content:
        example: 正文内容
        type: string
      title:
        example: 第一章 起始
        type: string
    type: object
  service.ChapterHistoryResponse:
    properties:
      content:
        example: 正文内容
        type: string
      id:
        example: 1
        type: integer
      title:
        example: 第一章 起始
        type: string
      updatedAt:
        example: "2020-1-1 00:00:00"
        type: string
      version:
        example: 1
        type: integer
    type: object
  service.ChapterNewRequest:
    properties:
      content:
        type: string
      subsectionId:
        type: integer
      title:
        type: string
      workId:
        type: integer
    required:
    - content
    - subsectionId
    - title
    - workId
    type: object
  service.ChapterSaveRequest:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
...
info:
  contact: {}
  description: 属于国内的同人圈
  license: {}
  termsOfService: https://github.com/shyptr/archiveofourown
  title: 同人圈
  version: "1.0"
paths:
...
  /api/v1/chapter:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.ChapterNewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 新建草稿章节
      tags:
      - 章节
  /api/v1/chapter/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 章节ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除章节
      tags:
      - 章节
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 章节ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.ChapterContentResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 查询章节内容
      tags:
      - 章节
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 章节ID
        in: path
        name: id
        required: true
        type: integer
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.ChapterSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 保存章节内容
      tags:
      - 章节
  /api/v1/chapter/{id}/{subsectionID}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 章节ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分卷ID
        in: path
        name: subsectionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 修改章节分卷信息
      tags:
      - 章节
  /api/v1/chapter/{id}/history:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 章节ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/service.ChapterHistoryResponse'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 查询章节历史版本
      tags:
      - 章节
...
  /api/v1/chapter/{id}/publish:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 章节ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 回收章节
      tags:
      - 章节
...
swagger: "2.0"